Robita/backend/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
fastapi==0.112.2
h11==0.14.0
idna==3.8
iso8601==1.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
pytz==2024.1
sniffio==1.3.1
starlette==0.38.2
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
uvicorn==0.30.6
